requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.23
WTForms==3.1.1
email-validator==2.1.0
cryptography==41.0.5
pycryptodome==3.19.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
gunicorn==21.2.0
python-jose==3.3.0
marshmallow==3.20.1
apispec==6.3.0
apispec-webframeworks==0.5.2 